requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
langchain==0.1.1
langchain-openai==0.1.0
langgraph==0.0.46
pydantic==2.5.0
PyPDF2==3.0.1
pdfplumber==0.10.3
numpy==1.24.3
faiss-cpu==1.7.4
scipy==1.11.4
python-multipart==0.0.6
